+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Allow setting a simple menu sidebar in the config
 * Add simple notification after adding/saving
 * Add page & window titles
+* Improve edit buttons
 
 ## 1.0.0 (2021-09-12)